Split rule to multiple files
