+++ lib/src/obsoleted/rules/member_ordering_extended/models/modifier.dart
@@ -1,0 +1,21 @@
+part of '../member_ordering_extended_rule.dart';
+
+@immutable
+class _Modifier {
+  final String type;
+
+  const _Modifier(this.type);
+
+  static const public = _Modifier('public');
+  static const private = _Modifier('private');
+  static const unset = _Modifier('unset');
+
+  static const all = [
+    public,
+    private,
+    unset,
+  ];
+
+  static _Modifier parse(String? name) => all
+      .firstWhere((type) => type.type == name, orElse: () => _Modifier.unset);
+}
